refactor(test): relocate pure database test slice into julie-core (own test binary)
Move remaining 4 sibling test files (37 tests) into crates/julie-core/src/tests/:
- vector_storage.rs (21 tests): stripped cfg(test) wrapper, rewrote crate::extractors -> julie_extractors
- memory_vectors.rs (10 tests): stripped cfg(test) wrapper, no import changes needed
- bulk_store_types_tests.rs (4 tests): rewrote crate::extractors -> julie_extractors
- bulk_store_types_tdd.rs (2 tests, was test_bulk_store_types.rs): same rewrites + added
  missing line_count field; renamed to avoid test_*.rs gitignore glob

Removes bulk_store_types_tests, memory_vectors, vector_storage from top-crate
src/tests/mod.rs core block. Total in julie-core: 107 tests, all database-layer.

Gates:
- cargo nextest run -p julie-core: 107/107 pass
- cargo nextest run --lib tests::core::database: 0 (moved to julie-core)
- cargo nextest run --lib tests::core::vector_storage: 0 (moved to julie-core)
- cargo build: green
