+++ .memories/2026-06-03/165831_3e11.md
@@ -1,0 +1,69 @@
+---
+id: checkpoint_3e118d62
+timestamp: 2026-06-03T16:58:31.581Z
+tags:
+  - julie-rescue
+  - crate-split
+  - task-7
+  - database-tests
+  - julie-core
+git:
+  branch: julie-rescue
+  commit: 35f16b60
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+    - crates/julie-core/src/tests/bulk_store_types_tests.rs
+    - crates/julie-core/src/tests/memory_vectors.rs
+    - crates/julie-core/src/tests/mod.rs
+    - crates/julie-core/src/tests/vector_storage.rs
+    - src/tests/core/bulk_store_types_tests.rs
+    - src/tests/core/memory_vectors.rs
+    - src/tests/core/vector_storage.rs
+    - src/tests/mod.rs
+summary: "Task 7 complete: relocate pure database test slice into julie-core"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - "cargo nextest run -p julie-core: 107/107 passed"
+  - "cargo nextest run --lib tests::core::database: 0 tests (zero from top
+    crate)"
+  - "cargo nextest run --lib tests::core::vector_storage: 0 tests (zero from top
+    crate)"
+  - "cargo build: Finished dev profile"
+next: "Commit as `refactor(test): relocate pure database test slice into
+  julie-core (own test binary)`, then report to team-lead for Task 8"
+---
+
+## Task 7 complete: relocate pure database test slice into julie-core
+
+### WHAT
+Moved the final 4 sibling test files into `crates/julie-core/src/tests/`:
+- `vector_storage.rs` (21 tests) — stripped `#[cfg(test)] mod tests {}` wrapper, rewrote `crate::extractors::SymbolKind` → `julie_extractors::SymbolKind`, split helpers import into local + `julie_test_support::symbol_builder`
+- `memory_vectors.rs` (10 tests) — stripped wrapper, no import changes needed
+- `bulk_store_types_tests.rs` (4 tests) — rewrote `crate::extractors::*` → `julie_extractors::*`
+- `test_bulk_store_types.rs` (2 tests) — rewrote `crate::extractors::*` → `julie_extractors::*`, `crate::extractors::Symbol {}` → `julie_extractors::Symbol {}`, added missing `line_count: 0` field to FileInfo struct literal in perf test
+
+### WHY
+Complete Task 7 of the Julie rescue crate-split. Database tests must live in `julie-core`'s own test binary so they relink only when `julie-core` changes, not when handler/tools code changes.
+
+### HOW
+- `crates/julie-core/src/tests/mod.rs`: added 4 new `mod` declarations
+- Removed `bulk_store_types_tests`, `memory_vectors`, `vector_storage` from `src/tests/mod.rs` core block
+- Deleted 4 source files from `src/tests/core/`
+- Note: `test_bulk_store_types` was never declared in `src/tests/mod.rs` (orphaned file)
+
+### IMPACT
+- `cargo nextest run -p julie-core`: **107/107 pass**
+- `cargo nextest run --lib tests::core::database`: **0 tests** (removed from top crate)
+- `cargo nextest run --lib tests::core::vector_storage`: **0 tests** (removed from top crate)
+- `cargo build`: **green**
+- Total test slice in julie-core: 107 tests, all database-layer, zero handler/tools deps
