+++ crates/julie-core/src/tests/mod.rs
@@ -3,3 +3,7 @@ mod database;
 mod database_init_race;
 mod database_lightweight_query;
 mod database_row_mapping;
+mod vector_storage;
+mod memory_vectors;
+mod bulk_store_types_tests;
+mod bulk_store_types_tdd;
